Refactor map popups
